fix: fixed typos && refactored check file path logic in lf-uss, lfds command && changed alias of local file uss file comapre command from lfuss to lf-uss
