+++ packages/cli/src/zosfiles/compare/Compare.definition.ts
@@ -13,7 +13,7 @@ import {ICommandDefinition} from "@zowe/imperative";
 import {DatasetDefinition} from "./ds/Dataset.definition";
 import { LocalfileDatasetDefinition } from "./lfds/LocalfileDataset.definition";
 import {UssFileDefinition} from './uss/UssFile.definition';
-import { LocalfileUssFileDefinition } from "./lfuss/LocalfileUss.definition";
+import { LocalfileUssFileDefinition } from "./lf-uss/LocalfileUss.definition";
 
 import i18nTypings from "../-strings-/en";
